Reestructuracion del proyecto en apps (todavia no funciona)
