+++ Clickrentacar/settings.py
@@ -29,13 +29,16 @@ ALLOWED_HOSTS = []
 # Application definition
 
 INSTALLED_APPS = [
-    'cracAPP.apps.CracappConfig',
     'django.contrib.admin',
     'django.contrib.auth',
     'django.contrib.contenttypes',
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
+    'core',
+    'alquiler',
+    'usuario',
+    'vehiculo'
 ]
 
 MIDDLEWARE = [
