+++ cracAPP/models.py
@@ -1,6 +1,1 @@
-from django.db import models
-
-from .models import ciudad, cuentaUsuario, cuentaAdministrador, denegacion, licencia, perfilAlquila, \
-    perfilPropietario, registroAlquiler, solicitudAlquiler, solicitudRegistro, vehiculo, vehiculoPapeles
-
 # Create your models here.
