--- cracAPP/models/__init__.py
@@ -1,2 +1,0 @@
-from . import ciudad, cuentaUsuario, cuentaAdministrador, denegacion, licencia, perfilAlquila, \
-    perfilPropietario, registroAlquiler, solicitudAlquiler, solicitudRegistro, vehiculo, vehiculoPapeles
