--- cracAPP/models/registroAlquiler.py
@@ -1,8 +1,0 @@
-from django.db import models
-
-
-class RegistroAlquiler(models.Model):
-    id = models.IntegerField(primary_key=True)
-    # fecha emision
-    # fecha devolucion
-    # solicitud alquiler
