--- cracAPP/models/vehiculoPapeles.py
@@ -1,8 +1,0 @@
-from django.db import models
-
-from . import vehiculo
-
-
-class VehiculoPapeles(models.Model):
-    id = models.IntegerField(primary_key=True)
-    vehiculo = models.ForeignKey(vehiculo.Vehiculo, on_delete=models.CASCADE)
