+++ vehiculo/admin.py
@@ -1,0 +1,7 @@
+from django.contrib import admin
+from vehiculo.models import Vehiculo
+from vehiculo.models import VehiculoPapeles
+
+# Register your models here.
+admin.site.register(Vehiculo)
+admin.site.register(VehiculoPapeles)
